+++ src/views/PatronsPreRegistrationListContainer/hooks/useNewRecordHandler.test.js
@@ -8,6 +8,7 @@ import {
 
 import useNewRecordHandler from './useNewRecordHandler';
 import useUserDuplicatesCheck from './useUserDuplicatesCheck';
+import useProcessPreRegisteredUser from './useProcessPreRegisteredUser';
 
 jest.mock('react-query', () => ({
   useMutation: jest.fn(),
@@ -19,17 +20,21 @@ jest.mock('react-router-dom', () => ({
 }));
 
 jest.mock('./useUserDuplicatesCheck', () => jest.fn());
+jest.mock('./useProcessPreRegisteredUser', () => jest.fn());
 
 describe('useNewRecordHandler', () => {
   let mockHistoryPush;
   let mockCheckDuplicates;
+  let mockHandlePreRegisteredUser;
 
   beforeEach(() => {
     mockHistoryPush = jest.fn();
     mockCheckDuplicates = jest.fn();
+    mockHandlePreRegisteredUser = jest.fn();
 
     useHistory.mockReturnValue({ push: mockHistoryPush });
     useUserDuplicatesCheck.mockReturnValue({ checkDuplicates: mockCheckDuplicates });
+    useProcessPreRegisteredUser.mockReturnValue({ handlePreRegisteredUser: mockHandlePreRegisteredUser });
 
     useMutation.mockImplementation(({ mutationFn, onSuccess }) => ({
       mutateAsync: async (user) => {
@@ -55,5 +60,16 @@ describe('useNewRecordHandler', () => {
       pathname: '/users/pre-registration-records/duplicates/userId',
       state: { search: '' },
     });
+  });
+
+  it('should create new user record when duplicates are not found', async () => {
+    mockCheckDuplicates.mockResolvedValue(false);
+    const { result } = renderHook(() => useNewRecordHandler());
+
+    await act(async () => {
+      await result.current.handle({ contactInfo: { email: 'test@example.com' } });
+    });
+
+    expect(mockHandlePreRegisteredUser).toHaveBeenCalled();
   });
 });